+++ src/lib.rs
@@ -26,7 +26,7 @@ pub mod field;
 pub mod kzg;
 pub mod polynomial;
 pub mod setup;
-pub mod tinyrsa;
+pub mod tiny_rsa;
 
 use core::{
   fmt::{self, Display, Formatter},
